Refactor rest proxy filter
