+++ src/main/java/io/antmedia/cluster/IClusterStore.java
@@ -10,6 +10,8 @@ public interface IClusterStore
 
 	public ClusterNode getClusterNode(String nodeId);
 	
+	public ClusterNode getClusterNodeFromIP(String ipAddress);
+	
 	public long getNodeCount();
 		
 	public boolean deleteNode(String nodeId);
