+++ src/test/java/io/antmedia/test/db/DBStoresUnitTest.java
@@ -330,7 +330,7 @@ public class DBStoresUnitTest {
 		assertEquals(0, vodList.size());
 		
 		for (int i = 0; i < 10; i++) {
-			dataStore.addVod(new VoD("stream", "111223" + (int)(Math.random() * 1000),  "path", "vod", 1517239808, 111, 17933, 1190525, VoD.STREAM_VOD, "1112233" + (int)(Math.random() * 91000), null));
+			assertNotNull(dataStore.addVod(new VoD("stream", "111223" + (int)(Math.random() * 100000),  "path", "vod", 1517239808, 111, 17933, 1190525, VoD.STREAM_VOD, "1112233" + (int)(Math.random() * 91000), null)));
 		}
 		
 		vodList = dataStore.getVodList(6, 4, null, null, null, null);
